Create Count Negative Numbers in a Sorted Matrix.cpp
since the rows and columns are sorted. The top-left element is greater than all its left and down elements and also every element that occurs in that region. So to reduce the time complexity, we need to prune. We do the processing row-wise, we start with the greater element of the smallest row.

In the While loop,
if the current is negative, all its left are negative as well. We check if the same can be said about its upper element in the above row.